+++ AppDesiginFormat/.idea/misc.xml
@@ -71,6 +71,10 @@
         <entry key="..\:/D/android-InnDiary/AppDesiginFormat/app/src/main/res/layout/fragment_diary_update.xml" value="0.36614583333333334" />
         <entry key="..\:/D/android-InnDiary/AppDesiginFormat/app/src/main/res/layout/fragment_setting.xml" value="0.33" />
         <entry key="..\:/D/android-InnDiary/AppDesiginFormat/app/src/main/res/layout/fragment_todo.xml" value="0.2740036231884058" />
+        <entry key="..\:/D/android-InnDiary/AppDesiginFormat/app/src/main/res/layout/popup_date.xml" value="0.36614583333333334" />
+        <entry key="..\:/D/android-InnDiary/AppDesiginFormat/app/src/main/res/layout/popup_delete.xml" value="0.36614583333333334" />
+        <entry key="..\:/D/android-InnDiary/AppDesiginFormat/app/src/main/res/layout/popup_login.xml" value="0.36614583333333334" />
+        <entry key="..\:/D/android-InnDiary/AppDesiginFormat/app/src/main/res/layout/popup_time.xml" value="0.36614583333333334" />
         <entry key="..\:/D/android-InnDiary/AppDesiginFormat/app/src/main/res/layout/setting_item.xml" value="0.2740036231884058" />
         <entry key="..\:/D/androidStudy/August/1st_week/AppDesiginFormat/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.249" />
         <entry key="..\:/D/androidStudy/August/1st_week/AppDesiginFormat/app/src/main/res/drawable/home_ic.xml" value="0.34" />
